backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
sqlalchemy==2.0.23
geoalchemy2==0.14.2
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
redis==5.0.1
rq==1.15.1
httpx==0.25.2
